snippet - unordered containers
Discovering Modern C++: An Intensive Course for Scientists, Engineers, and Programmers
Peter Gottschling

Hash tables are containers with very efficient search. In contrast to the preceding containers, hash maps have a constant complexity for a single access (when the hash function is reasonably good). To avoid name clashes with existing software, the standards committee refrained from names with “hash” and prepended “unordered” to the names of the ordered containers.

Unordered containers can be used in the same manner as their ordered counterparts - yielding the same results as map. If desired, a user-defined hash function can be provided.